Pull out the header and action bar into new components
add rootDomain to SingleAppApiData
